Fix circular model import issue
We can't grab direct references to the class constructors at module load
time, because (at load time) that reference will be undefined for
whatever module in the cycle is loaded second. Instead, just hold a
reference to the full module.exports object and grab the reference to
the class when needed during execution. We don't need this reference
during the module load, so we should always be safe here. Must also make
sure not to overwrite the exports object after requiring the other model
and instead append exports as properties to original empty
module.exports object. If we construct a new object and assign it to
module.exports, already loaded modules won't see the change.